refactor: Move Role context provider to pages.
