+++ src/admin/pages/Role/Context/type.ts
@@ -1,0 +1,6 @@
+import { Role } from '@shared/types';
+import { SWRResponse } from 'swr';
+
+export type RoleContext = {
+  getRoles: () => SWRResponse<Role[]>;
+};
